requirements.txt
agency-swarm[fastapi]>=1.2.0
openai-agents[litellm]>=1.2.0
fastapi
uvicorn
youtube-transcript-api
python-dotenv
notion-client
